+++ seafile/SeafStarredFile.m
@@ -19,8 +19,8 @@
 - (id)initWithConnection:(SeafConnection *)aConnection
                     repo:(NSString *)aRepo
                     path:(NSString *)aPath
-                   mtime:(int)mtime
-                    size:(int)size
+                   mtime:(long long)mtime
+                    size:(long long)size
                      org:(int)org
 {
     NSString *name = aPath.lastPathComponent;
